separate booking MVC first step and Fix worng loading Issue
